pyproject.toml
[project]
name = "backend"
version = "0"
description = ""
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
  "fastapi (>=0.115.12,<0.116.0)",
  "uvicorn[standard] (>=0.34.0,<0.35.0)",
  "python-dotenv (>=1.1.0,<2.0.0)",
  "slowapi (>=0.1.9,<0.2.0)",
  "atproto (>=0.0.59,<0.0.60)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
mypy = "^1.15.0"
